Finished cleaning up Authors module
