+++ content/all-contributors.json
@@ -1,7 +1,5 @@
 {
-  "files": [
-    "README.md"
-  ],
+  "files": ["README.md"],
   "imageSize": 100,
   "commit": false,
   "contributors": [
@@ -10,11 +8,7 @@
       "name": "Réjon Taylor-Foster",
       "avatar_url": "https://avatars1.githubusercontent.com/u/9157341?v=4",
       "profile": "https://rejontaylor.com",
-      "contributions": [
-        "code",
-        "design",
-        "example"
-      ]
+      "contributions": ["code", "design", "example"]
     }
   ],
   "contributorsPerLine": 7,
