+++ content/en/test.mdx
@@ -243,3 +243,11 @@
 </Box>
 
 </Authors>
+
+
+### Render a specifc author by their username
+
+<Authors author="MaximumCrash"/>
+
+**You can also use an @ for the author, but it's not necessary**
+<Authors list author="@MaximumCrash"/>
