+++ src/modules/authors/index.js
@@ -1,0 +1,5 @@
+export { default as GridCell } from "./GridCell";
+export { default as ListElement } from "./ListElement";
+export { default as MDXtoAuthor } from "./MDXtoAuthor";
+export { default as AuthorRenderer } from "./AuthorRenderer";
+export { default as allContributorKeys } from "./allContributorKeys";
