+++ src/modules/utility/index.js
@@ -1,4 +1,3 @@
 export { default as Video } from "./Video";
 export { default as Aligner } from "./Aligner";
 export { default as SEO } from "./seo.js";
-export { default as Authors } from "./Authors";
